requirements.txt
discord.py
python-dotenv
genshin
cryptography